refactor: Biographyをリファクタ
