+++ src/components/common/index.ts
@@ -1,2 +1,1 @@
-export * from './Biography'
 export * from './Navigation'
