+++ src/components/ui/Biography/index.ts
@@ -1,0 +1,1 @@
+export * from './Biography'
