+++ src/pages/works.tsx
@@ -2,7 +2,8 @@ import { VStack } from '@chakra-ui/react'
 import { GetStaticProps } from 'next'
 import React from 'react'
 import Parser from 'rss-parser'
-import { Biography, Navigation } from 'src/components/common'
+import { Navigation } from 'src/components/common'
+import { Biography } from 'src/components/ui/Biography'
 import { Layout } from 'src/components/ui/Layout'
 import { MetaTag } from 'src/components/ui/MetaTag'
 import {
